Carved out  all the node control functions and put them in their own class (NodeControl.cs)
